+++ EventListener/RepeatedFieldListener.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace APY\JsFormValidationBundle\Listener;
+namespace APY\JsFormValidationBundle\EventListener;
 
 use APY\JsFormValidationBundle\Generator\PostProcessEvent;
 
@@ -17,15 +17,15 @@ class RepeatedFieldListener
 {
     public function onJsfvPostProcess(PostProcessEvent $event)
     {
-        $formFields = $event->getFormView()->getChildren();
+        $formFields = $event->getFormView()->children;
         $fieldsConstraints = $event->getFieldsConstraints();
 
         foreach ($formFields as $formField) {
-            if (in_array('repeated', $formField->get('types'))) {
-                $formFieldId = $formField->get('id');
+            if (isset($formField->vars['type']) && $formField->vars['type'] == 'repeated') {
+                $formFieldId = $formField->vars['id'];
 
                 // Get the real fields name of the repeated type form
-                $repeatedNames = array_keys($formField->get('value'));
+                $repeatedNames = array_keys($formField->vars['value']);
                 $formFieldId_first = $formFieldId.'_'.$repeatedNames[0];
                 $formFieldId_second = $formFieldId.'_'.$repeatedNames[1];
 
@@ -41,14 +41,18 @@ class RepeatedFieldListener
                 }
 
                 // Get invalid message from the form
-                $invalid_message = $formField->get('invalid_message');
-                foreach ($formField->get('invalid_message_parameters') as $invalid_message_parameter => $value){
-                    $invalid_message = str_replace($invalid_message_parameter,'{{ '.$invalid_message_parameter.' }}',$invalid_message);
+                $invalid_message = isset($formField->vars['invalid_message']) ? $formField->vars['invalid_message'] : null;
+                $invalid_message_parameters = isset($formField->vars['invalid_message_parameters']) ?
+                    $formField->vars['invalid_message_parameters'] : null;
+                if (!empty($invalid_message_parameters)) {
+                    foreach ($invalid_message_parameters as $invalid_message_parameter => $value){
+                        $invalid_message = str_replace($invalid_message_parameter,'{{ '.$invalid_message_parameter.' }}',$invalid_message);
+                    }
                 }
-                $invalid_message_parameters = json_encode($formField->get('invalid_message_parameters'));
+                $invalid_message_parameters = json_encode($invalid_message_parameters);
 
                 $fieldsConstraints->addFieldConstraint($formFieldId_second, array(
-                    'name' => 'Repeated',
+                    'name'       => 'Repeated',
                     'parameters' => "{first_name: '$formFieldId_first', second_name: '$formFieldId_second', invalid_message: '$invalid_message', invalid_message_parameters: $invalid_message_parameters}"
                 ));
             }